Update of list
"List" and "Header" are both coliding with eachother and shouldn't be used together in the form of the list, either choose one or the other. Before the change every additional party/contributor to the list, started with "1.", taking out the "<h3>" tag, the list is now shown how it should be, in a numerical order.